Add get_process_name
Adds a LUA function get_process_name implemented in c_get_process_name that retrieves the binary name of the process of the window by its PID which is retrieved through wnck_window_get_pid.
It uses a shell command "ps c %d | tail -n 1 | awk '{print $5}'" where %d is the PID.
A better approach would be using "/proc/%d/cmdline" however this includes the whole path and all command arguments which would first have to be cleaned up to be usable, however my C skills are too low to do this.
Some windows, like the Steam Friends window, return 0 as PID so this is not a complete solution.

The reason this function is useful is because many programs change their window name and application name, like Discord (depending on the chat that is open) or Spotify (depending on the active song).
Other programs also report "Unnamed window" as their window and application name, like Spotify upon startup, making it impossible to correctly identify a program just by their window or application name.

Again, my C skills suck, so please feel free to improve this functionality!