Add set_adjust_for_decoration([bool])
Allow for situations where moving or resizing the window is done
incorrectly, i.e.
   set_window_position(0,0)
results in the window decoration being taken into account twice, i.e.
the window (including decoration) is offset from the top left corner by
the width of the left side decoration and the height of the title bar.

This is currently off by default.