crash and memory leak when changing config files
When editing config files while dp2 is running, the application frequently
crashes. It is best reproducible when (using vim as an editor works great,
it creates temporary files which trigger this) opening devilspie2.lua and
just write the file. It triggers many reload events, at some point it
crashes.

The reason for this is that the list structure is not correctly cleaned up
in clear_file_lists() causing a crash if the function is called again before
the lists are repopulated. Even if it does not crash memory is lost since
the old list is not properly freed.