Allow rules to be run when a window name/title changes
The Existing 3 event types aren't quite enough if, for example you have multiple browser windows open, and want them to be on different workspaces.  At the point the windows are opened they will have a generic name.
Attaching a listener for "name-changed" to each window as they are opened allows these changes to be caught (and acted upon) reliably.